Added genres and actors tables (models) , added relation to movies table, added more attributes to movies, executed migrations to db

+++ Backend/models/ratings.js
@@ -25,7 +25,7 @@ module.exports = function (sequelize, DataTypes) {
         },
       },
       rating: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(3, 1),
         allowNull: false,
         validate: {
           min: 1,
